@@ -8,7 +8,26 @@
 import SwiftUI
 
 struct OnboardingName: View {
-    @State private var name: String = ""
+    @AppStorage("pet_name") var pet_name: String = ""
+    @AppStorage("onboard") var onboard: Bool = true
+    
+    func saveName(){
+        Task {
+            do {
+                let currentUser = try await supabase.auth.session.user
+                
+                let result = try await supabase.database
+                        .from("pet_profiles")
+                        .update([
+                            "name": pet_name
+                        ])
+                        .eq("owner", value: currentUser.id)
+                        .execute()
+            } catch{
+                debugPrint(error)
+            }
+        }
+    }
     var body: some View {
             VStack {
                 
@@ -28,14 +47,20 @@ struct OnboardingName: View {
                         .padding()
                         .foregroundColor(Color.white)
                     
+                
                     
-                    
-                TextField("Name", text: $name)
+                TextField("Name", text: $pet_name)
                     .padding()
                     .background(Color(.secondarySystemBackground))
                     .cornerRadius(5.0)
                     .frame(maxWidth: .maximum(80, 300))
                     .padding(.bottom, 20)
+                
+                if !onboard {
+                    Button(action: {saveName()}){
+                        Text("Save").foregroundColor(.white)
+                    }
+                }
             }
         .background(Color("bubbles-background"))
     
